Minor improvements

+++ src/event_handler.ts
@@ -34,7 +34,7 @@ export function handleChangeEvent(changeEvent: vscode.TextDocumentChangeEvent) {
 		console.log(`Skipping: No repoPath`);
 		return; 
 	}
-	const shadowPath = `${SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${relPath}`;
+	const shadowPath = path.join(SHADOW_REPO, `${repoPath}/${branch}/${relPath}`);
 	const shadowExists = fs.existsSync(shadowPath);
 	if (!shadowExists) { 
 		// TODO: Create shadow file?
@@ -79,27 +79,30 @@ export function handleFilesCreated(changeEvent: vscode.FileCreateEvent) {
 	changeEvent.files.forEach((file) => {
 		const filePath = file.path;
 		// Skip for directory
-		if (fs.lstatSync(filePath).isDirectory()) { return; }
+		const lstat = fs.lstatSync(filePath);
+		if (lstat.isDirectory()) { return; }
 		const relPath = filePath.split(`${repoPath}/`)[1];
 		// Skip .git/ and syncignore files
 		if (shouldIgnoreFile(repoPath, relPath)) { return; }
+		console.log(`FileCreated: ${filePath}`);
 		const branch = getBranchName({ altPath: repoPath }) || DEFAULT_BRANCH;
-		const destOriginals = `${ORIGINALS_REPO}/${repoPath.slice(1)}/${branch}/${relPath}`;
-		const destOriginalsPathSplit = destOriginals.split("/");
-		const destOriginalsBasePath = destOriginalsPathSplit.slice(0, destOriginalsPathSplit.length-1).join("/");
-		const destShadow = `${SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${relPath}`;
+		const destShadow = path.join(SHADOW_REPO, `${repoPath}/${branch}/${relPath}`);
 		const destShadowPathSplit = destShadow.split("/");
 		const destShadowBasePath = destShadowPathSplit.slice(0, destShadowPathSplit.length-1).join("/");
-		if (fs.existsSync(destOriginals)) { return; }
-		console.log(`FileCreated: ${filePath}`);
-		// Add file in originals repo
-		fs.mkdirSync(destOriginalsBasePath, { recursive: true });
-		// File destination will be created or overwritten by default.
-		fs.copyFileSync(filePath, destOriginals);
+
+		const destOriginals = path.join(ORIGINALS_REPO, `${repoPath}/${branch}/${relPath}`);
+		const destOriginalsPathSplit = destOriginals.split("/");
+		const destOriginalsBasePath = destOriginalsPathSplit.slice(0, destOriginalsPathSplit.length-1).join("/");
+
+		if (fs.existsSync(destShadow) || fs.existsSync(destOriginals)) { return; }
 		// Add file in shadow repo
 		fs.mkdirSync(destShadowBasePath, { recursive: true });
 		// File destination will be created or overwritten by default.
 		fs.copyFileSync(filePath, destShadow);  
+		// Add file in originals repo
+		fs.mkdirSync(destOriginalsBasePath, { recursive: true });
+		// File destination will be created or overwritten by default.
+		fs.copyFileSync(filePath, destOriginals);
 		// Add new diff in the buffer
 		manageDiff(repoPath, branch, relPath, "", true);
 	});
@@ -200,8 +203,8 @@ export function handleFilesRenamed(changeEvent: vscode.FileRenameEvent) {
 function handleRename(repoPath: string, branch: string, oldAbsPath: string, newAbsPath: string, isFile: boolean) {
 	const oldRelPath = oldAbsPath.split(`${repoPath}/`)[1];
 	const newRelPath = newAbsPath.split(`${repoPath}/`)[1];
-	const oldShadowPath = `${SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${oldRelPath}`;
-	const newShadowPath = `${SHADOW_REPO}/${repoPath.slice(1)}/${branch}/${newRelPath}`;
+	const oldShadowPath = path.join(SHADOW_REPO, `${repoPath}/${branch}/${oldRelPath}`);
+	const newShadowPath = path.join(SHADOW_REPO, `${repoPath}/${branch}/${newRelPath}`);
 
 	// rename file in shadow repo
 	fs.renameSync(oldShadowPath, newShadowPath);
